+++ PyNFSe/base/assinatura.py
