+++ PyNFSe/base/models.py
@@ -1,5 +1,4 @@
 import typing
-
 from datetime import datetime
 from decimal import Decimal
